test: split the L2 profile status tests into their own module
The architecture gate's `new-test-size` check failed: adding the
readiness-vs-measurement coverage took tests/test_l2_real_profiles.py to
1214 lines, over the 1200-line test maximum.

Split by responsibility rather than trimmed, per AGENTS.md — the way to
shrink a file is to give part of it an owner. test_l2_real_profiles.py
keeps the profile definitions, validation, and prepare-script rendering;
the new test_l2_real_profile_status.py owns the three classes that answer
a different question (can a status claim completed work that never
happened). The importlib load of the script and the operand fixture move
to a shared tests/_l2_real_profiles_support.py, so both modules share one
loaded module object and a monkeypatch written against one works in the
other.

No test is removed or weakened: 178 tests before, 178 after.
